Improve code stype
